Refactor: libpacemaker: separate remote-related code into own file
There are no code changes other than commenting, formatting, and renaming
apply_remote_node_ordering() to pcmk__order_remote_connection_actions().
